// src/redux/interfaces/IUser.ts
export interface IUser {
    avatar_code: string | null;
    blocked: null;
    board_game_player: boolean;
    email: string | null;
    enough_questions: boolean;
    name: string;
    only_chat_with_friends: boolean;
    q_reviewer: number;
    qc: boolean;
    show_gift: boolean;
    user_id: string;
}